+++ src/types/UserProjectData.ts
@@ -4,7 +4,10 @@ export interface UserProjectData {
     org : string,
     name : string,
     owner? : string,
-    guidelines : string,
+    description? : string,
+    title?: string,
+    // guidelines are a keyed list of guidelines for the project
+    guidelines? : Record<string, string>[],
     resources : ProjectResource[],
     lastUpdated : number,
 }